Split input files and rename "RequiredInput" to "Input"
